+++ SharpSource/SharpSource/Utilities/DiagnosticId.cs
@@ -46,4 +46,5 @@ public static class DiagnosticId
     public const string StaticInitializerAccessedBeforeInitialization = "SS045";
     public const string UnboundedStackalloc = "SS046";
     public const string LinqTraversalBeforeFilter = "SS047";
+    public const string LockingOnDiscouragedObject = "SS048";
 }